Move isTraySupported () into separate file
Addresses comment https://github.com/Zettlr/Zettlr/pull/2026#discussion_r645694707
